refactor(docx-core): split document_view into focused modules
document_view.ts had grown to ~1,580 lines mixing several independent
concerns. Split it along natural seams into sibling modules:
- document_view-headings.ts  (heading detection/heuristics + types)
- document_view-styles.ts     (formatting fingerprinting + style discovery)
- document_view-comments.ts   (inline comment-marker injection + TOON tokenizing)
- document_view-toon.ts        (TOON rendering)
document_view.ts retains the orchestration core (buildDocumentView /
buildNodesForDocumentView / footnote + hyperlink helpers) and re-exports the
moved symbols.

Why: improves navigability/reviewability of a large core file and sets up
cleaner boundaries for the upcoming ODF document-view work.

Behavior-preserving: code was moved, not rewritten; the public export surface
of primitives/document_view.js is byte-identical before/after (verified by
diffing the runtime export list), so no importer changes. Reuses the shared
xml-helpers; no inline defensive idioms reintroduced. Not Lean-coupled
(document_view is neither referenced by verification/lean nor imported by the
Lean differential tests).

Verified: docx-core build + 1302 vitest tests pass; full pre-submit chain
(build, lint:workspaces, test:run, spec-coverage, conformance citations + doc)
green.

Peer review (Gemini + Codex) pending.

Fixes: #315
