+++ packages/docx-core/src/primitives/document_view.ts
@@ -8,16 +8,34 @@ import { HIGHLIGHT_TAG } from './semantic_tags.js';
 import { type AnnotatedRun, type FormattingBaseline, type FormattingMode, computeModalBaseline, computeParagraphFontBaseline, emitFormattingTags, mergeAdjacentTags } from './formatting_tags.js';
 import type { RelsMap } from './relationships.js';
 import { isReservedFootnote } from './footnotes.js';
+import {
+  deriveHeading,
+  detectRunInHeader,
+  detectTitleCapsCentered,
+  extractHeaderInfo,
+  suppressSignatureClusters,
+  type HeaderFormatting,
+  type HeadingValue,
+  type HeuristicHeadingSource,
+} from './document_view-headings.js';
+import { discoverStyles, fingerprintKey, type DocumentStyles, type FormattingFingerprint } from './document_view-styles.js';
+import { findTaggedTextInsertionIndex, type DocumentViewComment } from './document_view-comments.js';
 
-const SHORT_HEADER_MAX_LENGTH = 50;
-const MAX_HEADER_TEXT_LENGTH = 60;
-// Centered ALL-CAPS titles (e.g. NVCA COI's `AMENDED AND RESTATED CERTIFICATE
-// OF INCORPORATION OF FOO INC.`) routinely exceed 60 chars in real corporate
-// documents. The 60-char cap on `extractHeaderInfo` exists to avoid emitting a
-// "leading words = header" guess from long body prose, which doesn't apply to
-// the standalone-title detector.
-const MAX_CENTERED_TITLE_LENGTH = 120;
-const STYLE_EXAMPLE_TEXT_PREVIEW_LENGTH = 50;
+export type { HeaderFormatting, HeadingSource, HeadingValue, HeuristicHeadingSource } from './document_view-headings.js';
+export { discoverStyles } from './document_view-styles.js';
+export type { DocumentStyleInfo, DocumentStyles, FormattingFingerprint } from './document_view-styles.js';
+export { INLINE_COMMENT_MARKER_RUNTIME, TOON_INLINE_TAG_RE, collectInlineCommentMarkers, tokenizeToonInline } from './document_view-comments.js';
+export type { DocumentViewComment, DocumentViewCommentRange, ToonCommentMarker, ToonCommentMarkerMap, ToonInlineToken } from './document_view-comments.js';
+export {
+  collectTableMarkerInfo,
+  formatTableMarker,
+  formatToonCommentEndnoteLines,
+  formatToonCommentLines,
+  formatToonCommentsEndnotesBlock,
+  formatToonDataLine,
+  renderToon,
+  renderToonWithCommentEndnotes,
+} from './document_view-toon.js';
 
 function getWAttr(el: Element, localName: string): string | null {
   return getAttributeSafe(el, OOXML.W_NS, localName, 'w');
@@ -56,36 +74,6 @@ function emitHighlightTagsFromParagraph(p: Element): string {
   return out.join('');
 }
 
-export type HeaderFormatting = {
-  bold: boolean;
-  italic: boolean;
-  underline: boolean;
-};
-
-export type HeadingSource =
-  | 'word_style'
-  | 'run_in_header'
-  | 'title_with_period'
-  | 'title_with_colon'
-  | 'title_caps_centered'
-  | 'title_bare';
-
-export type HeuristicHeadingSource = Exclude<HeadingSource, 'word_style'>;
-
-export type HeadingValue = {
-  /**
-   * Heading label text. Semantics depend on `source`:
-   * - `word_style`: the full paragraph text (the entire paragraph IS the heading).
-   * - All heuristic sources (`run_in_header`, `title_with_period`, `title_with_colon`,
-   *   `title_caps_centered`, `title_bare`): only the extracted heading prefix.
-   *   For example, on `"Indemnification. The Company shall …"` the value is
-   *   `"Indemnification"`, not the whole paragraph.
-   */
-  text: string;
-  source: HeadingSource;
-  level: number | null;
-};
-
 export type ListMetadata = {
   list_level: number; // -1 for non-list
   label_type: LabelType | null;
@@ -94,29 +82,6 @@ export type ListMetadata = {
   header_style: HeuristicHeadingSource | null;
   header_formatting: HeaderFormatting | null;
   is_auto_numbered: boolean;
-};
-
-export type FormattingFingerprint = {
-  list_level: number;
-  left_indent_pt: number;
-  first_line_indent_pt: number;
-  style_name: string;
-  alignment: ParagraphAlignment;
-};
-
-export type DocumentStyleInfo = {
-  style_id: string;
-  display_name: string;
-  fingerprint: FormattingFingerprint;
-  example_node_id: string;
-  example_text: string;
-  count: number;
-  dominant_alignment: ParagraphAlignment;
-};
-
-export type DocumentStyles = {
-  styles: Map<string, DocumentStyleInfo>;
-  fingerprint_to_style: Map<string, string>; // fingerprintKey -> style_id
 };
 
 export type TableContext = {
@@ -131,44 +96,6 @@ export type TableContext = {
   para_in_cell: number;     // 0-based paragraph index within cell
   cell_para_count: number;  // Total paragraphs in this cell
 };
-
-export type DocumentViewCommentRange = {
-  startParagraphId: string;
-  endParagraphId: string;
-  startRunIndex?: number;
-  startCharOffset?: number;
-  endRunIndex?: number;
-  endCharOffset?: number;
-};
-
-export type DocumentViewComment = {
-  id: number;
-  author: string;
-  date: string | null;
-  initials: string;
-  text: string;
-  replies: DocumentViewComment[];
-  range?: DocumentViewCommentRange;
-};
-
-export const INLINE_COMMENT_MARKER_RUNTIME = Symbol('inline_comment_marker_runtime');
-
-type InlineCommentMarkerRuntime = {
-  startVisibleOffset: number;
-  endVisibleOffset: number;
-  suppressInlineMarkers: boolean;
-};
-
-type DocumentViewCommentWithRuntime = DocumentViewComment & {
-  [INLINE_COMMENT_MARKER_RUNTIME]?: InlineCommentMarkerRuntime;
-};
-
-export type ToonCommentMarker = {
-  offset: number;
-  marker: string;
-};
-
-export type ToonCommentMarkerMap = Map<string, ToonCommentMarker[]>;
 
 export type DocumentViewNode = {
   id: string; // _bk_*
@@ -205,812 +132,6 @@ export type DocumentViewNode = {
    */
   visible_offset_correction?: number;
 };
-
-function fingerprintKey(fp: FormattingFingerprint): string {
-  // Stable JSON-ish key used for Map lookups.
-  return `${fp.list_level}|${fp.left_indent_pt.toFixed(1)}|${fp.first_line_indent_pt.toFixed(1)}|${fp.style_name}|${fp.alignment}`;
-}
-
-/**
- * v0.3: Compact style fingerprint token.
- * Concatenates style name, list level, alignment, and indentation for token-efficient LLM context.
- * Example: "Normal:L-1:LEFT:I0:H0"
- */
-function computeFingerprintToken(fp: FormattingFingerprint, styleId?: string): string {
-  const name = styleId || fp.style_name || 'body';
-  const level = `L${fp.list_level}`;
-  const align = fp.alignment;
-  const indent = `I${Math.round(fp.left_indent_pt)}`;
-  const hanging = `H${Math.round(fp.first_line_indent_pt)}`;
-  return `${name}:${level}:${align}:${indent}:${hanging}`;
-}
-
-// Pattern-based header detection fallback (ported from Python ingestor._extract_header_info).
-const HEADER_PATTERN = /^([A-Z][^.!?:]*(?:\s+[A-Z][^.!?:]*)*)([.:]?)(?:\s|$)/;
-
-function extractHeaderInfo(cleanText: string): { header_text: string | null; header_style: HeuristicHeadingSource | null } {
-  if (!cleanText || cleanText.length < 2) return { header_text: null, header_style: null };
-  if (!/^[A-Z]/.test(cleanText)) return { header_text: null, header_style: null };
-
-  const stripped = cleanText.trim();
-  if (stripped.length <= SHORT_HEADER_MAX_LENGTH) {
-    if (stripped.endsWith('.')) return { header_text: stripped.slice(0, -1), header_style: 'title_with_period' };
-    if (stripped.endsWith(':')) return { header_text: stripped.slice(0, -1), header_style: 'title_with_colon' };
-
-    const words = stripped.split(/\s+/);
-    if (words.length <= 5) return { header_text: stripped, header_style: 'title_bare' };
-    return { header_text: null, header_style: null };
-  }
-
-  const m = HEADER_PATTERN.exec(stripped);
-  if (!m) return { header_text: null, header_style: null };
-  const headerText = (m[1] ?? '').trim();
-  const terminator = m[2] ?? '';
-  const remaining = stripped.slice(m[0].length);
-  if (!remaining || headerText.length > MAX_HEADER_TEXT_LENGTH) return { header_text: null, header_style: null };
-
-  if (terminator === '.') return { header_text: headerText, header_style: 'title_with_period' };
-  if (terminator === ':') return { header_text: headerText, header_style: 'title_with_colon' };
-  // Long-paragraph regex matches without an explicit terminator are body prose
-  // (e.g. "Termination of Section 2.2(d)(i) shall not affect ..."), not headers.
-  // Bare titles only fire from the short-paragraph branch above.
-  return { header_text: null, header_style: null };
-}
-
-function deriveHeading(
-  paragraphStyleId: string | null,
-  cleanText: string,
-  headerText: string | null,
-  headerStyle: HeuristicHeadingSource | null,
-  isInTableCell: boolean,
-): HeadingValue | undefined {
-  const styleMatch = paragraphStyleId ? /^Heading([1-6])$/.exec(paragraphStyleId) : null;
-  if (styleMatch) {
-    return {
-      text: cleanText,
-      source: 'word_style',
-      level: Number.parseInt(styleMatch[1]!, 10),
-    };
-  }
-
-  // Inside table cells, heuristic detectors (run_in_header, title_with_period,
-  // title_with_colon, title_bare) routinely fire on ordinary label/value content
-  // — "Name", "Purchase Price:", "Name: Acme" — which are not structural document
-  // headings. We keep the per-detector explanation on list_metadata.header_style
-  // for debugging, but suppress heuristic promotion into the canonical heading
-  // predicate. Word built-in heading styles inside cells remain real headings.
-  if (isInTableCell) return undefined;
-
-  if (headerText && headerStyle) {
-    return {
-      text: headerText,
-      source: headerStyle,
-      level: null,
-    };
-  }
-
-  return undefined;
-}
-
-function detectRunInHeader(params: {
-  paragraph: Element;
-  paragraphPPr: Element | null;
-  paragraphStyleId: string | null;
-  styles: StylesModel;
-}): { raw_text: string; formatting: HeaderFormatting; headerCharCount: number } | null {
-  const { paragraph, paragraphPPr, paragraphStyleId, styles } = params;
-  const punct = new Set(['.', ':', '-']);
-
-  // Use visible runs only (field code text stripped in getParagraphRuns()).
-  const runs = getParagraphRuns(paragraph);
-  if (runs.length === 0) return null;
-
-  // Group by run element, preserving order.
-  const orderedUniqueRuns: Element[] = [];
-  const seen = new Set<Element>();
-  for (const tr of runs) {
-    if (!seen.has(tr.r)) {
-      seen.add(tr.r);
-      orderedUniqueRuns.push(tr.r);
-    }
-  }
-
-  // Walk runs once, splitting into bold/underline header-prefix text and
-  // everything-after body text. The header → body transition is what
-  // distinguishes a run-in header (bold prefix + body) from a fully-bold
-  // signature label or defined-term lead-in.
-  let headerText = '';
-  let bodyText = '';
-  let formatting: HeaderFormatting | null = null;
-  let headerCharCount = 0;
-  let inHeader = true;
-
-  for (const r of orderedUniqueRuns) {
-    const fmt = extractEffectiveRunFormatting({ run: r, paragraphPPr, paragraphStyleId, styles });
-    const isHeaderStyle = fmt.bold || fmt.underline;
-    const ts = Array.from(r.getElementsByTagNameNS(OOXML.W_NS, W.t));
-    let runText = '';
-    for (const t of ts) runText += t.textContent ?? '';
-
-    if (inHeader && isHeaderStyle) {
-      headerText += runText;
-      headerCharCount += runText.length;
-      if (!formatting) formatting = { bold: fmt.bold, italic: fmt.italic, underline: fmt.underline };
-    } else {
-      inHeader = false;
-      bodyText += runText;
-    }
-  }
-
-  const trimmed = headerText.trim();
-  if (!trimmed) return null;
-  if (!punct.has(trimmed[trimmed.length - 1]!)) return null;
-  if (!formatting) return null;
-  // Require a real header-prefix → body transition: there must be non-whitespace
-  // body text after the bold/underline prefix. Trailing-whitespace-only "body"
-  // (e.g. a single bold run followed by a non-bold run that holds just `" "`)
-  // is not a transition — those are still whole-paragraph bold blocks
-  // (signature labels, all-bold short titles, etc.) and must be rejected.
-  if (!bodyText.trim()) return null;
-
-  return { raw_text: trimmed, formatting, headerCharCount };
-}
-
-/**
- * Detect a centered, ALL-CAPS, bold standalone title (e.g. an NVCA SPA's
- * `SERIES […] PREFERRED STOCK PURCHASE AGREEMENT` title).
- *
- * Strict gates only — fires only when the paragraph cannot be confused with
- * body prose, a placeholder, or a signature line:
- *   - paragraph alignment is CENTER
- *   - clean text contains no lowercase letters
- *   - clean text contains ≥ 3 ASCII letters AND ≥ 2 whitespace-separated
- *     word-tokens (so single-token bracketed placeholders like `[COMPANY]`
- *     and underscore-only signature lines like `____________` are rejected)
- *   - clean text is non-empty and ≤ MAX_CENTERED_TITLE_LENGTH
- *   - all visible runs are bold (a single non-bold char disqualifies)
- */
-function detectTitleCapsCentered(params: {
-  paragraph: Element;
-  paragraphPPr: Element | null;
-  paragraphStyleId: string | null;
-  alignment: ParagraphAlignment;
-  cleanTextNoLabel: string;
-  styles: StylesModel;
-}): { raw_text: string; formatting: HeaderFormatting } | null {
-  const { paragraph, paragraphPPr, paragraphStyleId, alignment, cleanTextNoLabel, styles } = params;
-  if (alignment !== 'CENTER') return null;
-  const trimmed = cleanTextNoLabel.trim();
-  if (!trimmed) return null;
-  if (trimmed.length > MAX_CENTERED_TITLE_LENGTH) return null;
-  if (/[a-z]/.test(trimmed)) return null;
-  // Content gate: punctuation/underscore-only signature lines and bracketed
-  // single-token placeholders (`[COMPANY]`, `[___]`, `<NAME>`) must not
-  // classify as titles. Real titles are multi-word ALL-CAPS phrases.
-  const letterCount = (trimmed.match(/[A-Z]/g) ?? []).length;
-  if (letterCount < 3) return null;
-  const wordTokens = trimmed.split(/\s+/).filter((w) => /[A-Z]/.test(w));
-  if (wordTokens.length < 2) return null;
-
-  const runs = getParagraphRuns(paragraph);
-  if (runs.length === 0) return null;
-  const orderedUniqueRuns: Element[] = [];
-  const seen = new Set<Element>();
-  for (const tr of runs) {
-    if (!seen.has(tr.r)) {
-      seen.add(tr.r);
-      orderedUniqueRuns.push(tr.r);
-    }
-  }
-
-  let formatting: HeaderFormatting | null = null;
-  let sawAnyText = false;
-  for (const r of orderedUniqueRuns) {
-    const ts = Array.from(r.getElementsByTagNameNS(OOXML.W_NS, W.t));
-    let runHasText = false;
-    for (const t of ts) {
-      if ((t.textContent ?? '').length > 0) {
-        runHasText = true;
-        break;
-      }
-    }
-    if (!runHasText) continue;
-    const fmt = extractEffectiveRunFormatting({ run: r, paragraphPPr, paragraphStyleId, styles });
-    if (!fmt.bold) return null;
-    sawAnyText = true;
-    if (!formatting) formatting = { bold: fmt.bold, italic: fmt.italic, underline: fmt.underline };
-  }
-  if (!sawAnyText || !formatting) return null;
-
-  return { raw_text: trimmed, formatting };
-}
-
-const SIGNATURE_LABEL_LINE_RE = /^[A-Z][a-zA-Z ]{0,28}:\s*$/;
-const SIGNATURE_LABEL_PREFIX_RE = /^[A-Z]+(?::\s|$)/;
-
-function isSignatureClusterLabel(text: string): boolean {
-  const trimmed = text.trim();
-  if (!trimmed) return false;
-  return SIGNATURE_LABEL_LINE_RE.test(trimmed) || SIGNATURE_LABEL_PREFIX_RE.test(trimmed);
-}
-
-function suppressSignatureClusters(nodes: DocumentViewNode[]): void {
-  if (nodes.length < 4) return;
-
-  const prefixMatches = new Array<number>(nodes.length + 1).fill(0);
-  for (let idx = 0; idx < nodes.length; idx++) {
-    prefixMatches[idx + 1] = prefixMatches[idx]! + (isSignatureClusterLabel(nodes[idx]!.clean_text) ? 1 : 0);
-  }
-
-  const coverage = new Array<number>(nodes.length + 1).fill(0);
-  for (let start = 0; start <= nodes.length - 4; start++) {
-    for (let end = start + 3; end < nodes.length; end++) {
-      const runLength = end - start + 1;
-      const matchCount = prefixMatches[end + 1]! - prefixMatches[start]!;
-      if ((matchCount * 4) < (runLength * 3)) continue;
-      coverage[start]! += 1;
-      coverage[end + 1]! -= 1;
-    }
-  }
-
-  let activeClusters = 0;
-  for (let idx = 0; idx < nodes.length; idx++) {
-    activeClusters += coverage[idx]!;
-    if (activeClusters <= 0) continue;
-
-    const node = nodes[idx]!;
-    // The density gate authorizes us to clear *labels* inside the window;
-    // non-label neighbors (real headings, body text) keep their detected
-    // heading metadata regardless of paragraph style. This avoids erasing
-    // an adjacent section heading or body line that happens to fall inside
-    // a window meeting the density threshold.
-    if (!isSignatureClusterLabel(node.clean_text)) continue;
-    node.header = '';
-    node.header_formatting = null;
-    node.list_metadata.header_text = null;
-    node.list_metadata.header_style = null;
-    node.list_metadata.header_formatting = null;
-  }
-}
-
-function inferSemanticName(params: {
-  fp: FormattingFingerprint;
-  nodes: DocumentViewNode[];
-}): { base_id: string; display_name: string } {
-  const { fp, nodes } = params;
-
-  // Find first label_type if present.
-  let labelType: LabelType | null = null;
-  for (const n of nodes) {
-    if (n.list_metadata.label_type) {
-      labelType = n.list_metadata.label_type;
-      break;
-    }
-  }
-
-  const listLevel = fp.list_level;
-
-  if (listLevel >= 0) {
-    if (listLevel === 0) {
-      if (labelType === LabelType.ARTICLE) return { base_id: 'article', display_name: 'Article Heading' };
-      if (labelType === LabelType.SECTION) return { base_id: 'section', display_name: 'Section Heading' };
-      if (labelType === LabelType.ROMAN) return { base_id: 'roman_section', display_name: 'Roman Numeral Section' };
-      return { base_id: 'top_level', display_name: 'Top-Level List Item' };
-    }
-    if (listLevel === 1) {
-      if (labelType === LabelType.LETTER) return { base_id: 'subsection', display_name: 'Subsection (a)/(A)' };
-      if (labelType === LabelType.NUMBER) return { base_id: 'subsection_number', display_name: 'Numbered Subsection' };
-      if (labelType === LabelType.ROMAN) return { base_id: 'subsection_roman', display_name: 'Roman Subsection' };
-      return { base_id: 'level_1', display_name: `Level ${listLevel} List Item` };
-    }
-    if (labelType === LabelType.ROMAN) return { base_id: `level_${listLevel}_roman`, display_name: `Level ${listLevel} Roman` };
-    if (labelType === LabelType.LETTER) return { base_id: `level_${listLevel}_letter`, display_name: `Level ${listLevel} Letter` };
-    return { base_id: `level_${listLevel}`, display_name: `Level ${listLevel} List Item` };
-  }
-
-  // Non-list.
-  const styleName = fp.style_name.toLowerCase().replace(/\s+/g, '_');
-  if (fp.left_indent_pt > 0) return { base_id: 'indent_block', display_name: 'Indented Block' };
-  if (styleName.includes('heading') || styleName.includes('title')) return { base_id: 'heading', display_name: 'Heading' };
-  if (styleName.includes('quote') || styleName.includes('block')) return { base_id: 'block_quote', display_name: 'Block Quote' };
-  return { base_id: 'body', display_name: 'Body Text' };
-}
-
-export function discoverStyles(nodes: DocumentViewNode[]): DocumentStyles {
-  const groups = new Map<string, { fp: FormattingFingerprint; nodes: DocumentViewNode[] }>();
-  for (const n of nodes) {
-    const key = fingerprintKey(n.style_fingerprint);
-    const g = groups.get(key);
-    if (g) g.nodes.push(n);
-    else groups.set(key, { fp: n.style_fingerprint, nodes: [n] });
-  }
-
-  const used: Record<string, number> = {};
-  const styles = new Map<string, DocumentStyleInfo>();
-  const fpToStyle = new Map<string, string>();
-
-  for (const [fpKey, g] of groups.entries()) {
-    const { base_id, display_name } = inferSemanticName({ fp: g.fp, nodes: g.nodes });
-    let styleId = base_id;
-    if (used[base_id] !== undefined) {
-      used[base_id] += 1;
-      styleId = `${base_id}_${used[base_id]}`;
-    } else {
-      used[base_id] = 0;
-    }
-
-    const median = g.nodes[Math.floor(g.nodes.length / 2)]!;
-    const info: DocumentStyleInfo = {
-      style_id: styleId,
-      display_name,
-      fingerprint: g.fp,
-      example_node_id: median.id,
-      example_text: median.clean_text.slice(0, STYLE_EXAMPLE_TEXT_PREVIEW_LENGTH),
-      count: g.nodes.length,
-      dominant_alignment: g.fp.alignment,
-    };
-    styles.set(styleId, info);
-    fpToStyle.set(fpKey, styleId);
-  }
-
-  return { styles, fingerprint_to_style: fpToStyle };
-}
-
-function headerStripFromText(params: { header: string; text: string }): string {
-  // Mirrors Python TOONRenderer header stripping.
-  const { header } = params;
-  let { text } = params;
-  if (!header) return text;
-
-  const headerNorm = header.trim().toLowerCase();
-  const textLower = text.toLowerCase();
-
-  for (const punct of [':', '.', '-', ';', ''] as const) {
-    const testPrefix = `${headerNorm}${punct}`;
-    if (textLower.startsWith(testPrefix)) {
-      text = text.slice(testPrefix.length).trimStart();
-      return text;
-    }
-  }
-
-  if (text.startsWith(header)) {
-    text = text.slice(header.length).replace(/^[.:\-;]+/, '').trimStart();
-  }
-  return text;
-}
-
-// Matches the exact set of TOON inline formatting tags that emitFormattingTags() can emit:
-//   <b>, </b>, <i>, </i>, <u>, </u>, <highlight>, </highlight>,
-//   <a href="...">, </a>, <font ATTR=...>, </font>
-// Anything else in the form `<...>` is literal document text (e.g., `<Borrower>` placeholders
-// in legal templates, or stylesheet samples like `<font>`) and must be counted as visible
-// characters, not skipped as markup.
-//
-// Note the opening `a`/`font` alternative requires `\s[^>]*` (mandatory attributes), because
-// the formatter only emits `<a href="...">` and `<font ATTR=...>` — never bare `<a>` or
-// `<font>`. Allowing the bare forms would cause literal `<a>` / `<font>` in document text to
-// be silently skipped, shifting marker positions.
-export const TOON_INLINE_TAG_RE = /^(?:<\/?(?:b|i|u|highlight)>|<\/(?:a|font)>|<(?:a|font)\s[^>]*>)/;
-
-/** A single token produced by {@link tokenizeToonInline}. */
-export type ToonInlineToken =
-  | { kind: 'tag'; value: string }
-  | { kind: 'text'; value: string };
-
-/**
- * Split a TOON inline-tag string (`DocumentViewNode.tagged_text` produced with
- * `show_formatting`) into an ordered list of `tag` and `text` tokens, using the exact same
- * grammar (`TOON_INLINE_TAG_RE`) the formatter emits. Consecutive literal characters are
- * coalesced into one `text` token. This is the shared tokenization primitive used by
- * downstream serializers (Markdown today, HTML next) so they never reason about the tag
- * grammar independently and drift from the emitter.
- */
-export function tokenizeToonInline(text: string): ToonInlineToken[] {
-  const tokens: ToonInlineToken[] = [];
-  let buffer = '';
-  for (let i = 0; i < text.length; i++) {
-    const tagLen = toonTagLengthAt(text, i);
-    if (tagLen > 0) {
-      if (buffer) {
-        tokens.push({ kind: 'text', value: buffer });
-        buffer = '';
-      }
-      tokens.push({ kind: 'tag', value: text.slice(i, i + tagLen) });
-      i += tagLen - 1;
-      continue;
-    }
-    buffer += text[i];
-  }
-  if (buffer) tokens.push({ kind: 'text', value: buffer });
-  return tokens;
-}
-
-function toonTagLengthAt(text: string, i: number): number {
-  if (text[i] !== '<') return 0;
-  const match = TOON_INLINE_TAG_RE.exec(text.slice(i));
-  return match ? match[0].length : 0;
-}
-
-function countVisibleTextCharacters(text: string): number {
-  let visibleCount = 0;
-  for (let i = 0; i < text.length; i++) {
-    const tagLen = toonTagLengthAt(text, i);
-    if (tagLen > 0) {
-      i += tagLen - 1;
-      continue;
-    }
-    visibleCount++;
-  }
-  return visibleCount;
-}
-
-function findTaggedTextInsertionIndex(text: string, visibleOffset: number): number {
-  if (visibleOffset <= 0) return 0;
-
-  let visibleCount = 0;
-  for (let i = 0; i < text.length; i++) {
-    if (visibleCount === visibleOffset) return i;
-
-    const tagLen = toonTagLengthAt(text, i);
-    if (tagLen > 0) {
-      i += tagLen - 1;
-      continue;
-    }
-
-    visibleCount++;
-  }
-
-  return text.length;
-}
-
-function injectToonCommentMarkers(
-  text: string,
-  markers: readonly ToonCommentMarker[],
-): string {
-  if (markers.length === 0) return text;
-
-  let result = text;
-  for (const { offset, marker } of markers) {
-    const insertionIndex = findTaggedTextInsertionIndex(result, offset);
-    result = result.slice(0, insertionIndex) + marker + result.slice(insertionIndex);
-  }
-  return result;
-}
-
-type InlineCommentMarkerCandidate = {
-  id: number;
-  startParagraphId: string;
-  endParagraphId: string;
-  startParagraphIndex: number;
-  startOffset: number;
-  endOffset: number;
-};
-
-type InlineCommentMarkerGroup = {
-  closes: InlineCommentMarkerCandidate[];
-  opens: InlineCommentMarkerCandidate[];
-};
-
-function collectInlineCommentMarkerCandidates(
-  comments: readonly DocumentViewComment[],
-  paragraphIndexById: ReadonlyMap<string, number>,
-  candidates: InlineCommentMarkerCandidate[],
-): void {
-  for (const comment of comments) {
-    const runtime = (comment as DocumentViewCommentWithRuntime)[INLINE_COMMENT_MARKER_RUNTIME];
-    if (comment.range && runtime && !runtime.suppressInlineMarkers) {
-      candidates.push({
-        id: comment.id,
-        startParagraphId: comment.range.startParagraphId,
-        endParagraphId: comment.range.endParagraphId,
-        startParagraphIndex: paragraphIndexById.get(comment.range.startParagraphId) ?? Number.MAX_SAFE_INTEGER,
-        startOffset: runtime.startVisibleOffset,
-        endOffset: runtime.endVisibleOffset,
-      });
-    }
-
-    if (comment.replies.length > 0) {
-      collectInlineCommentMarkerCandidates(comment.replies, paragraphIndexById, candidates);
-    }
-  }
-}
-
-function compareInlineCommentCloseOrder(
-  left: InlineCommentMarkerCandidate,
-  right: InlineCommentMarkerCandidate,
-): number {
-  if (left.startParagraphIndex !== right.startParagraphIndex) {
-    return right.startParagraphIndex - left.startParagraphIndex;
-  }
-  if (left.startOffset !== right.startOffset) {
-    return right.startOffset - left.startOffset;
-  }
-  return right.id - left.id;
-}
-
-export function collectInlineCommentMarkers(
-  nodes: readonly DocumentViewNode[],
-): ToonCommentMarkerMap {
-  const paragraphIndexById = new Map<string, number>();
-  for (let index = 0; index < nodes.length; index++) {
-    paragraphIndexById.set(nodes[index]!.id, index);
-  }
-
-  const candidates: InlineCommentMarkerCandidate[] = [];
-  for (const node of nodes) {
-    if (node.comments && node.comments.length > 0) {
-      collectInlineCommentMarkerCandidates(node.comments, paragraphIndexById, candidates);
-    }
-  }
-
-  const groupedByParagraph = new Map<string, Map<number, InlineCommentMarkerGroup>>();
-  for (const candidate of candidates) {
-    const startOffsets = groupedByParagraph.get(candidate.startParagraphId) ?? new Map<number, InlineCommentMarkerGroup>();
-    const startGroup = startOffsets.get(candidate.startOffset) ?? { closes: [], opens: [] };
-    startGroup.opens.push(candidate);
-    startOffsets.set(candidate.startOffset, startGroup);
-    groupedByParagraph.set(candidate.startParagraphId, startOffsets);
-
-    const endOffsets = groupedByParagraph.get(candidate.endParagraphId) ?? new Map<number, InlineCommentMarkerGroup>();
-    const endGroup = endOffsets.get(candidate.endOffset) ?? { closes: [], opens: [] };
-    endGroup.closes.push(candidate);
-    endOffsets.set(candidate.endOffset, endGroup);
-    groupedByParagraph.set(candidate.endParagraphId, endOffsets);
-  }
-
-  const markersByParagraph = new Map<string, ToonCommentMarker[]>();
-  for (const [paragraphId, offsetGroups] of groupedByParagraph.entries()) {
-    const markers: ToonCommentMarker[] = [];
-    const sortedOffsets = Array.from(offsetGroups.keys()).sort((left, right) => right - left);
-    for (const offset of sortedOffsets) {
-      const group = offsetGroups.get(offset);
-      if (!group) continue;
-
-      const closes = [...group.closes].sort(compareInlineCommentCloseOrder);
-      const opens = [...group.opens].sort((left, right) => left.id - right.id);
-      const marker =
-        closes.map((comment) => `[cm-end:${comment.id}]`).join('') +
-        opens.map((comment) => `[cm-start:${comment.id}]`).join('');
-      if (!marker) continue;
-      markers.push({ offset, marker });
-    }
-
-    if (markers.length > 0) {
-      markersByParagraph.set(paragraphId, markers);
-    }
-  }
-
-  return markersByParagraph;
-}
-
-/**
- * Format a single toon data line for one DocumentViewNode.
- * Handles table-context-aware style (th/td) and header stripping.
- */
-export function formatToonDataLine(
-  n: DocumentViewNode,
-  options?: { compact?: boolean; commentMarkers?: ToonCommentMarkerMap },
-): string {
-  let text = n.tagged_text;
-  let header = n.header;
-  let strippedPrefixVisibleLength = 0;
-
-  if (header) {
-    const strippedText = headerStripFromText({ header, text });
-    strippedPrefixVisibleLength = Math.max(
-      0,
-      countVisibleTextCharacters(text) - countVisibleTextCharacters(strippedText),
-    );
-    text = strippedText;
-  }
-  if (header && !text) {
-    text = header;
-    header = '';
-    strippedPrefixVisibleLength = 0;
-  }
-
-  const commentMarkers = options?.commentMarkers?.get(n.id);
-  if (commentMarkers && commentMarkers.length > 0) {
-    // Comment marker offsets are computed against the FULL paragraph visible text (raw
-    // run/char counting in `getComments()`). To translate to `tagged_text` positions we
-    // subtract:
-    //  1. `visible_offset_correction` — chars stripped at build time when extracting the
-    //     manual list label and trimming following whitespace.
-    //  2. `strippedPrefixVisibleLength` — chars stripped at format time by the run-in-header
-    //     extraction above.
-    const totalCorrection = (n.visible_offset_correction ?? 0) + strippedPrefixVisibleLength;
-    text = injectToonCommentMarkers(
-      text,
-      commentMarkers.map(({ offset, marker }) => ({
-        offset: Math.max(0, offset - totalCorrection),
-        marker,
-      })),
-    );
-  }
-
-  const tc = n.table_context;
-  let style: string;
-  if (tc) {
-    style = tc.is_header_row
-      ? `th(${tc.row_index},${tc.col_index})`
-      : `td(${tc.row_index},${tc.col_index})`;
-  } else {
-    style = options?.compact
-      ? computeFingerprintToken(n.style_fingerprint, n.style)
-      : n.style;
-  }
-  return `${n.id} | ${n.list_label} | ${header} | ${style} | ${text}`;
-}
-
-/**
- * Collect table marker info (dimensions) from nodes for #TABLE markers.
- * Column headers are NOT included in the marker — they appear once in the th() rows.
- */
-export function collectTableMarkerInfo(
-  nodes: readonly Pick<DocumentViewNode, 'table_context'>[],
-): Map<number, { id: string; totalRows: number; totalCols: number }> {
-  const info = new Map<number, { id: string; totalRows: number; totalCols: number }>();
-  for (const n of nodes) {
-    const tc = n.table_context;
-    if (!tc) continue;
-    if (!info.has(tc.table_index)) {
-      info.set(tc.table_index, {
-        id: tc.table_id,
-        totalRows: tc.total_rows,
-        totalCols: tc.total_cols,
-      });
-    }
-  }
-  return info;
-}
-
-/**
- * Format a #TABLE marker line from collected table info.
- * Headers are omitted — they appear exactly once in the th(0,N) data rows.
- */
-export function formatTableMarker(info: { id: string; totalRows: number; totalCols: number }): string {
-  return `#TABLE ${info.id} | ${info.totalRows} rows × ${info.totalCols} cols`;
-}
-
-function escapeToonCommentField(value: string): string {
-  return value
-    .replaceAll('\r\n', '\\n')
-    .replaceAll('\r', '\\r')
-    .replaceAll('\n', '\\n')
-    .replaceAll('|', '\\|');
-}
-
-function formatCommentDate(date: string | null): string {
-  return date ?? '-';
-}
-
-function collectToonCommentLines(
-  comment: DocumentViewComment,
-  paragraphId: string,
-  parentId?: number,
-): string[] {
-  const author = escapeToonCommentField(comment.author || '-');
-  const date = formatCommentDate(comment.date);
-  const text = escapeToonCommentField(comment.text);
-  const line = parentId == null
-    ? `#COMMENT ${paragraphId} c${comment.id} ${author} ${date} | ${text}`
-    : `#REPLY c${comment.id} -> c${parentId} ${author} ${date} | ${text}`;
-
-  return [
-    line,
-    ...comment.replies.flatMap((reply) => collectToonCommentLines(reply, paragraphId, comment.id)),
-  ];
-}
-
-export function formatToonCommentLines(node: Pick<DocumentViewNode, 'id' | 'comments'>): string[] {
-  return node.comments?.flatMap((comment) => collectToonCommentLines(comment, node.id)) ?? [];
-}
-
-function collectToonCommentEndnoteLines(
-  comment: DocumentViewComment,
-  paragraphId: string,
-  parentId?: number,
-): string[] {
-  const author = escapeToonCommentField(comment.author || '-');
-  const date = formatCommentDate(comment.date);
-  const text = escapeToonCommentField(comment.text);
-  const line = parentId == null
-    ? `c${comment.id} @ ${paragraphId} ${author} ${date} | ${text}`
-    : `c${comment.id} -> c${parentId} ${author} ${date} | ${text}`;
-
-  return [
-    line,
-    ...comment.replies.flatMap((reply) => collectToonCommentEndnoteLines(reply, paragraphId, comment.id)),
-  ];
-}
-
-export function formatToonCommentEndnoteLines(node: Pick<DocumentViewNode, 'id' | 'comments'>): string[] {
-  return node.comments?.flatMap((comment) => collectToonCommentEndnoteLines(comment, node.id)) ?? [];
-}
-
-export function formatToonCommentsEndnotesBlock(
-  nodes: readonly Pick<DocumentViewNode, 'id' | 'comments'>[],
-): string[] {
-  const commentLines = nodes.flatMap((node) => formatToonCommentEndnoteLines(node));
-  return commentLines.length > 0
-    ? ['#COMMENTS', ...commentLines]
-    : [];
-}
-
-export function renderToon(nodes: DocumentViewNode[], options: { compact?: boolean } = {}): string {
-  const lines: string[] = ['#SCHEMA id | list_label | header | style | text'];
-  const commentMarkers = collectInlineCommentMarkers(nodes);
-  const lineOptions = { ...options, commentMarkers };
-
-  // Pre-scan: collect table marker info for #TABLE lines
-  const tableInfo = collectTableMarkerInfo(nodes);
-
-  let currentTableIndex: number | null = null;
-
-  for (const n of nodes) {
-    const tc = n.table_context;
-    const nodeTableIndex = tc ? tc.table_index : null;
-
-    // Close previous table if we left it or moved to a different table
-    if (currentTableIndex !== null && nodeTableIndex !== currentTableIndex) {
-      lines.push('#END_TABLE');
-      currentTableIndex = null;
-    }
-
-    // Open new table if entering one
-    if (nodeTableIndex !== null && currentTableIndex === null) {
-      const info = tableInfo.get(nodeTableIndex);
-      if (info) lines.push(formatTableMarker(info));
-      currentTableIndex = nodeTableIndex;
-    }
-
-    lines.push(formatToonDataLine(n, lineOptions));
-    lines.push(...formatToonCommentLines(n));
-  }
-
-  // Close any open table at end
-  if (currentTableIndex !== null) {
-    lines.push('#END_TABLE');
-  }
-
-  return lines.join('\n');
-}
-
-export function renderToonWithCommentEndnotes(
-  nodes: DocumentViewNode[],
-  options: { compact?: boolean } = {},
-): string {
-  const lines: string[] = ['#SCHEMA id | list_label | header | style | text'];
-  const tableInfo = collectTableMarkerInfo(nodes);
-
-  let currentTableIndex: number | null = null;
-
-  for (const n of nodes) {
-    const tc = n.table_context;
-    const nodeTableIndex = tc ? tc.table_index : null;
-
-    if (currentTableIndex !== null && nodeTableIndex !== currentTableIndex) {
-      lines.push('#END_TABLE');
-      currentTableIndex = null;
-    }
-
-    if (nodeTableIndex !== null && currentTableIndex === null) {
-      const info = tableInfo.get(nodeTableIndex);
-      if (info) lines.push(formatTableMarker(info));
-      currentTableIndex = nodeTableIndex;
-    }
-
-    lines.push(formatToonDataLine(n, options));
-  }
-
-  if (currentTableIndex !== null) {
-    lines.push('#END_TABLE');
-  }
-
-  lines.push(...formatToonCommentsEndnotesBlock(nodes));
-
-  return lines.join('\n');
-}
 
 export type BuildDocumentViewOptions = {
   include_semantic_tags?: boolean;
